+++ spyder/utils/ipython/style.py
@@ -1,0 +1,192 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright © Spyder Project Contributors
+# Licensed under the terms of the MIT License
+# (see spyder/__init__.py for details)
+
+"""
+Style for IPython Console
+"""
+
+# Local imports
+from spyder.config.gui import get_color_scheme
+from spyder.config.main import CONF
+
+def get_style_sheet():
+    """Returns a CSS stylesheet with spyder color scheme settings.
+
+    The stylesheet can contain classes for:
+        Qt: QPlainTextEdit, QFrame, QWidget, etc
+       Pygments: .c, .k, .o, etc. (see PygmentsHighlighter)
+        IPython: .error, .in-prompt, .out-prompt, etc
+    """
+
+    def give_font_weight(is_bold):
+        if is_bold:
+            return 'bold'
+        else:
+            return 'normal'
+
+    def give_font_style(is_italic):
+        if is_italic:
+            return 'italic'
+        else:
+            return 'normal'
+
+    color_scheme = get_color_scheme(CONF.get('color_schemes', 'selected'))
+
+    fon_c, fon_fw, fon_fs = color_scheme['normal']
+    font_color =  fon_c
+    font_font_weight = give_font_weight(fon_fw)
+    font_font_style = give_font_style(fon_fs)
+    background_color = color_scheme['background']
+    selection_background_color = '#ccc'
+    error_color = 'red'
+    instance_color = color_scheme['instance'][0]
+    in_prompt_color = instance_color
+    in_prompt_number_font_weight = 'bold'
+    out_prompt_color = instance_color
+    out_prompt_number_font_weight = 'bold'
+    inverted_background_color = font_color
+    inverted_font_color = background_color
+    unmatched_parenthesis_color = color_scheme['unmatched_p']
+    matched_parenthesis_color = color_scheme['matched_p']
+
+    sheet = """QPlainTextEdit, QTextEdit {{
+                                          color: {} ;
+                                          font-weight: {};
+                                          font-style: {};
+                                          background-color: {};
+                                          selection-background-color: {}
+                                         }}
+              .error {{ color: {}; }}
+              .in-prompt {{ color: {}; }}
+              .in-prompt-number {{ font-weight: {}; }}
+              .out-prompt {{ color: {}; }}
+              .out-prompt-number {{ font-weight: {}; }}
+              /* .inverted is used to highlight selected completion */
+              .inverted {{ color: {}; background-color: {}; }}
+              """
+
+    sheet_formatted = sheet.format(font_color, font_font_weight,
+                                   font_font_style, background_color,
+                                   selection_background_color,
+                                   error_color,
+                                   in_prompt_color,
+                                   in_prompt_number_font_weight,
+                                   out_prompt_color,
+                                   out_prompt_number_font_weight,
+                                   inverted_background_color,
+                                   inverted_font_color)
+
+    return sheet_formatted
+
+def get_syntax_style():
+    """Create a .py with the spyder custom style as a Pygment class.
+
+    The file is stored in the pygment/styles folder.
+    Returns the name of this file (whithout path nor entension)
+    """
+
+    def give_font_weight(is_bold):
+        if is_bold:
+            return 'bold'
+        else:
+            return ''
+
+    def give_font_style(is_italic):
+        if is_italic:
+            return 'italic'
+        else:
+            return ''
+
+    selected_color_scheme = CONF.get('color_schemes', 'selected')
+    color_scheme = get_color_scheme(selected_color_scheme)
+
+    fon_c, fon_fw, fon_fs = color_scheme['normal']
+    font_color =  fon_c
+    font_font_weight = give_font_weight(fon_fw)
+    font_font_style = give_font_style(fon_fs)
+    key_c, key_fw, key_fs = color_scheme['keyword']
+    keyword_color =  key_c
+    keyword_font_weight = give_font_weight(key_fw)
+    keyword_font_style = give_font_style(key_fs)
+    bui_c, bui_fw, bui_fs = color_scheme['builtin']
+    builtin_color =  bui_c
+    builtin_font_weight = give_font_weight(bui_fw)
+    builtin_font_style = give_font_style(bui_fs)
+    str_c, str_fw, str_fs = color_scheme['string']
+    string_color =  str_c
+    string_font_weight = give_font_weight(str_fw)
+    string_font_style = give_font_style(str_fs)
+    num_c, num_fw, num_fs = color_scheme['number']
+    number_color =  num_c
+    number_font_weight = give_font_weight(num_fw)
+    number_font_style = give_font_style(num_fs)
+    com_c, com_fw, com_fs = color_scheme['comment']
+    comment_color =  com_c
+    comment_font_weight = give_font_weight(com_fw)
+    comment_font_style = give_font_style(com_fs)
+    def_c, def_fw, def_fs = color_scheme['definition']
+    definition_color =  def_c
+    definition_font_weight = give_font_weight(def_fw)
+    definition_font_style = give_font_style(def_fs)
+    ins_c, ins_fw, ins_fs = color_scheme['instance']
+    instance_color =  ins_c
+    instance_font_weight = give_font_weight(ins_fw)
+    instance_font_style = give_font_style(ins_fs)
+
+    import pygments
+    syntax_path = pygments.__file__.rpartition('/')[0] + '/styles/'
+    syntax_name = 'spyder'
+    syntax_filename = syntax_path + syntax_name + '.py'
+
+    syntax_style_imports = ['Name', 'Keyword', 'Comment', 'String',
+                            'Number', 'Punctuation']
+    imports_string = ", ".join(syntax_style_imports)
+
+    syntax_style_dic = {}
+    syntax_style_dic['Name'] = [font_font_style, font_font_weight,
+                                font_color]
+    syntax_style_dic['Name.Class'] = [definition_font_style,
+                                      definition_font_weight,
+                                      definition_color]
+    syntax_style_dic['Name.Function'] = [definition_font_style,
+                                         definition_font_weight,
+                                         definition_color]
+    syntax_style_dic['Name.Builtin'] = [builtin_font_style,
+                                        builtin_font_weight, builtin_color]
+    syntax_style_dic['Name.Builtin.Pseudo'] = [instance_font_style,
+                                               instance_font_weight,
+                                               instance_color]
+    syntax_style_dic['Keyword'] = [keyword_font_style, keyword_font_weight,
+                                   keyword_color]
+    syntax_style_dic['Keyword.Type'] = [builtin_font_style,
+                                        builtin_font_weight, builtin_color]
+    syntax_style_dic['Comment'] = [comment_font_style, comment_font_weight,
+                                   comment_color]
+    syntax_style_dic['String'] = [string_font_style, string_font_weight,
+                                  string_color]
+    syntax_style_dic['Number'] = [number_font_style, number_font_weight,
+                                  number_color]
+    syntax_style_dic['Punctuation'] = [font_font_style, font_font_weight,
+                                       font_color]
+
+    syntax_file_string = "# -*- coding: utf-8 -*-\n"
+    syntax_file_string += "from pygments.style import Style\n"
+    syntax_file_string += "from pygments.token import {}\n\n".format(imports_string)
+    syntax_file_string += "class SpyderStyle(Style):\n"
+    syntax_file_string += "\tstyles = {\n"
+    for key in syntax_style_dic:
+        syntax_file_string += "\t\t{}:'".format(key)
+        values = syntax_style_dic[key]
+        for val in values:
+            if val != '':
+                syntax_file_string += "{} ".format(val)
+        syntax_file_string += "',\n"
+    syntax_file_string += "\t}"
+
+    with open(syntax_filename, "w") as syntax_file:
+        syntax_file.write(syntax_file_string)
+
+    return syntax_name
